requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.7.5=h06a4308_1
cachetools=4.2.2=pypi_0
certifi=2021.5.30=py39h06a4308_0
charset-normalizer=2.0.4=pypi_0
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.10.0=pypi_0
cython=0.29.24=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.5=pypi_0
grpcio=1.39.0=pypi_0
idna=3.2=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=pypi_0
matplotlib=3.4.3=pypi_0
mkl=2021.3.0=h06a4308_520
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
numpy=1.21.2=pypi_0
oauthlib=3.1.1=pypi_0
olefile=0.46=py_0
opencv-python=4.5.3.56=pypi_0
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
pillow=8.3.1=py39h2c7a002_0
pip=21.2.4=py37h06a4308_0
protobuf=3.17.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pyparsing=2.4.7=pypi_0
python=3.9.6=h12debd9_1
python-dateutil=2.8.2=pypi_0
pytorch=1.9.0=py3.9_cuda11.1_cudnn8.0.5_0
pyyaml=5.4.1=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scipy=1.7.1=pypi_0
setuptools=52.0.0=py39h06a4308_0
six=1.16.0=pypi_0
sqlite=3.36.0=hc218d9a_0
tensorboard=2.6.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tk=8.6.10=hbc83047_0
torchvision=0.10.0=py39_cu111
tqdm=4.62.1=pypi_0
typing_extensions=3.10.0.0=pyh06a4308_0
tzdata=2021a=h5d7bf9c_0
urllib3=1.26.6=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.37.0=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
